elements {UpSetR}	R Documentation

Element query for queries parameter

Description

A query paramter to visualize specific elements of interest if queries = active (in custom plots only)

Usage

elements(func, query, ...)

Arguments

func
A functions provided internally

query
Input from the queries parameter assigned to the element function.

...
Additional parameters to be supplied internally

Note

See examples section of upset function on how to use this function in the queries parameter.

[Package UpSetR version 1.4.0 Index]