fromExpression {UpSetR}	R Documentation

Expression to UpSetR converters

Description

A function to convert an expression to a data frame compatible with UpSetR.

Usage

fromExpression(input)

Arguments

input
An vector (expression) to be converted to an input compatible with UpSetR

Note

See "Basic Usage" vignette for an example on how to use this function in UpSetR.

[Package UpSetR version 1.4.0 Index]