fromList {UpSetR}	R Documentation

List of named vectors to UpSetR converter

Description

A function to convert a list of named vectors to a data frame compatible with UpSetR.

Usage

fromList(input)

Arguments

input
A list of named vectors to be converted to a data frame compatible with UpSetR

Note

See "Basic Usage" vignette for an example on how to use this function in UpSetR.

[Package UpSetR version 1.4.0 Index]