intersects {UpSetR}	R Documentation

Intersection query for queries parameter

Description

A query paramter to visualize elements contained in specific intersections

Usage

intersects(func, query, ...)

Arguments

func
A functions provided internally

query
Input from the queries parameter assigned to the intersection function.

...
Additional parameters to be applied internally

Note

See examples section of upset function on how to use this function in the queries parameter.

[Package UpSetR version 1.4.0 Index]